改了一下版本提示机制，懒得和某些人解释

+++ PkpmLauncher/PkpmLauncher/BoundedQueue.h
@@ -110,7 +110,7 @@ namespace collection_utility
 		void deleteAt(const int index)
 		{
 			rangeNoError(index);
-			if((queue_.size()-1)==index)
+			if(static_cast<int>(queue_.size()-1)==index)
 				queue_.pop_back();
 			else
 				remove(index);
